+++ actionpack/lib/action_dispatch/system_test_case.rb
@@ -161,6 +161,14 @@ module ActionDispatch
       self.driver = SystemTesting::Driver.new(driver, **driver_options, &capabilities)
     end
 
+    # Configuration for the System Test application server
+    #
+    # By default this is localhost. This method allows the host and port to be specified manually.
+    def self.served_by(host:, port:)
+      Capybara.server_host = host
+      Capybara.server_port = port
+    end
+
     private
       def url_helpers
         @url_helpers ||=